Move IDocumentMappingService and dependencies from Razor.Workspaces to Remote.Razor
This includes moving one method in DelegatedCompletionHandler to remote, since it used the interface, and only had one caller, and combines all of the IDocumentMappingService extensions into one set
